Bug fixes and improvements

+++ include/stats.h
@@ -13,10 +13,36 @@ typedef struct {
     int products_counter;
 } cassa_log_t;
 
-//TODO fare tutto questa documentazione
-int log_client_stats(queue_t *client_log, int id, int prod, int time_instore, int time_inqueue, int queuecounter);
+/**
+ * Aggiorna il log del cliente con i parametri passati. Ritorna 0 in caso di successo, -1 altrimenti.
+ *
+ * @param client_log struttura dati del
+ * @param id identificatore univoco del cliente
+ * @param prod quanti prodotti ha acquistato il cliente
+ * @param time_instore quanto tempo il cliente ha passato all'interno del supermercato. Espresso in millisecondi
+ * @param time_inqueue quanto tempo il cliente ha passato in coda. Espresso in millisecondi
+ * @param queuecounter quante volte il cliente ha cambiato coda
+ * @return 0 in caso di successo, -1 altrimenti ed imposta errno.
+ */
+int log_client_stats(queue_t *client_log, int id, int prod, long time_instore, long time_inqueue, int queuecounter);
+
+/**
+ * Alloca la struttura dati del log di una cassa.
+ *
+ * @param id identificatore univoco della cassa
+ * @return il puntatore alla struttura dati allocata oppure NULL in caso di errore ed imposta errno
+ */
 cassa_log_t *alloc_cassa_log(size_t id);
+
+/**
+ * Dsistrugge la struttura dati del log della cassa passata per argomento.
+ *
+ * @param log puntatore alla struttura dati del log di una cassa
+ * @return 0 in caso di successo, -1 altrimenti ed imposta errno
+ */
 int destroy_cassa_log(cassa_log_t *log);
+
+//TODO finire questa documentazione
 int log_client_served(cassa_log_t *cassa_log, long time, int products);
 int log_cassa_opening_time(cassa_log_t *cassa_log, long time);
 void log_cassa_closed(cassa_log_t *cassa_log);
